% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{app}
\alias{app}
\title{Launch Shiny App}
\usage{
app(name = "template", ...)
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch Shiny App
}
