% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{clone}
\alias{clone}
\title{Clone a new shiny app from a template}
\usage{
clone(app = "basic_template", dir = app)
}
\arguments{
\item{app}{the template app to copy}

\item{dir}{the app directory to create}
}
\value{
boolean (if the directory was copied)
}
\description{
Clone a new shiny app from a template
}
