% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{newapp}
\alias{newapp}
\title{Set up a new shiny app from the template}
\usage{
newapp(dir = "myapp")
}
\arguments{
\item{dir}{the app directory to create}
}
\value{
boolean (if the directory was copied)
}
\description{
Set up a new shiny app from the template
}
